+++ pkg/logger/slog/service.go
@@ -38,16 +38,20 @@ import (
 
 type slogFabric struct {
 	zapLogMakerSvc zapLogEntryService
-	e              errorFormatterService
+
+	e errorFormatterService
 }
 
 func (s *slogFabric) NewLoggerEntry(name string,
 	fields ...any,
 ) *slog.Logger {
-	return slog.New(newZapHandler(nil, s.e, s.zapLogMakerSvc.NewLoggerEntry(name, fields...)))
+	return slog.New(newZapHandler(nil, s.e,
+		s.zapLogMakerSvc.NewLoggerEntry(name, fields...)))
 }
 
-func NewSLogMaker(zapLogMakerSvc zapLogEntryService, errFmtSvc errorFormatterService) *slogFabric {
+func NewSLogMaker(zapLogMakerSvc zapLogEntryService,
+	errFmtSvc errorFormatterService,
+) *slogFabric {
 	return &slogFabric{
 		zapLogMakerSvc: zapLogMakerSvc,
 		e:              errFmtSvc,