+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added golangci linter settings
 ### Changed
 * Cleaned up code by linter rules
+* Removed support of options flow from zap.Logger implementation of slog.Handler
 
 ## [v0.0.7] - 25.09.2024
 ### Added